Sync user state across tabs (#45)
Adds basic synchronization of user state across tabs when an update has written to local storage. This is to improve the experience for users using multiple tabs and to prevent onboarding progress that have been made in one tab from being overwritten by another. Implementation is simple and perhaps a bit naive but I believe it should cover most (if not all) scenarios we care about.